Split amp-sidebar component files
