+++ extensions/amp-sidebar/1.0/base-element.js
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {CSS as COMPONENT_CSS} from './component.jss';
+import {PreactBaseElement} from '../../../src/preact/base-element';
+import {Sidebar} from './component';
+
+export class BaseElement extends PreactBaseElement {}
+
+/** @override */
+BaseElement['Component'] = Sidebar;
+
+/** @override */
+BaseElement['passthrough'] = true;
+
+/** @override */
+BaseElement['shadowCss'] = COMPONENT_CSS;
+
+/** @override */
+BaseElement['props'] = {
+  'side': {attr: 'side', type: 'string'},
+};
