+++ extensions/amp-sidebar/1.0/component.jss.js
