+++ extensions/amp-sidebar/1.0/storybook/Basic.js
@@ -15,7 +15,7 @@
  */
 
 import * as Preact from '../../../../src/preact';
-import {Sidebar} from '../sidebar';
+import {Sidebar} from '../component';
 import {boolean, color, select, withKnobs} from '@storybook/addon-knobs';
 import {withA11y} from '@storybook/addon-a11y';
